+++ src/main/java/com/follydev/gestiondestock/services/impl/CommandeClientServiceImpl.java
@@ -245,6 +245,12 @@ public class CommandeClientServiceImpl implements CommandeClientService {
             log.error("Commande client ID is NULL");
             return;
         }
+        List<LigneCommandeClient> ligneCommandeClients = ligneCommandeClientRepository.findAllByCommandeClientId(id);
+        if(!ligneCommandeClients.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimer cette commande car elle rattachée à un client",
+                    ErrorCodes.COMMANDE_CLIENT_ALREADY_IN_USE);
+        }
+
         commandeClientRepository.deleteById(id);
     }
